backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
paho-mqtt==1.6.1
pika==1.3.2
cohere==4.37
openai==1.3.7
pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
wordcloud==1.9.3
